backend/requirements.txt
flask
jsonify
requests
load_dotenv